feat: modify existing rows instead of skipping
Github may give partial data when polling in the
middle of the day. With the skipping mechanism,
we never correct this partial data. Now we update
all historical data that we have access to each
time we poll.

Marginally higher load on the DB, but trivially
so. Opens up possibility for more frequent polling
and inner-day updates.